separate out bearing and speed functions
